Drop duration from `files.log` in tests
While this field should depend only on network time a user reported
seeing values differing from the baseline. I imagine that this could
happen due to very heavy system lead which caused us to switch to system
time. Since this value is not interesting for the test anyway drop it
from the baseline.

Closes #9.